✨ Feat:#141-챌린지 표 컴포넌트 연결
